# Replace manual-iteration with Iterators.zip() for faster copyto!