--- a/base/abstractarray.jl
+++ b/base/abstractarray.jl
@@ -979,11 +979,8 @@ function copyto_unaliased!(deststyle::IndexStyle, dest::AbstractArray, srcstyle:
             end
         else
             # Dual-iterator implementation
-            ret = iterate(iterdest)
-            @inbounds for a in src
-                idx, state = ret
-                dest[idx] = a
-                ret = iterate(iterdest, state)
+            for (idx, idy) in Iterators.zip(iterdest, itersrc)
+                @inbounds dest[idx] = src[idy]
             end
         end
     end