@N5N3 2021-01-22
It's hard to understand that replacing `zip(iterdest,itersrc)` with `zip(itersrc,iterdest)` can also accelerate it on my PC.
```julia
function zip_copyto!(dest, src)
    iterdest, itersrc = eachindex(dest), eachindex(src)
    for (idx, idy) in zip(iterdest, itersrc)
        @inbounds dest[idx] = src[idy]
    end
end
function zip_rev_copyto!(dest, src)
    iterdest, itersrc = eachindex(dest), eachindex(src)
    for (idy, idx) in zip(itersrc, iterdest)
        @inbounds dest[idx] = src[idy]
    end
end
a = randn(1000,1000)
a_ = view(a,1:600,:)
b = randn(500,1000)

@btime base_copyto!(a_,b) #   705.800 μs (0 allocations: 0 bytes)
@btime zip_copyto!(a_,b)  #   337.600 μs (0 allocations: 0 bytes)
@btime zip_rev_copyto!(a_,b)  #  280.099 μs (0 allocations: 0 bytes)
```