@kimikage 2021-02-19
> It's hard to understand that replacing zip(iterdest,itersrc) with zip(itersrc,iterdest) can also accelerate it on my PC.

In this example, `iterdest` is `CartesianIndices` and `itersrc` is `OneTo`. When we use `CartesianIndex` to access the memory, it should be converted to a linear index. The conversion involves multiplications, but it can be simplified to just additions. In a two-dimensional case like this one, for example, the iteration loop can be decomposed into the column loop and the row loop, and it is very similar to what we are doing in `iterate(iterdest::CartesianIndices, state)`.

Thus, the key point is whether or not the compiler is aware of this optimization. I'm just guessing here, but it would probably be more compiler friendly to do `iterate(iterdest, state)` later, since there is no `iterate(ititersrc, state) === nothing` branch between it and `dest[idx]`.

Therefore, whether `zip(itersrc, iterdest)` or `zip(iterdest, itersrc)` is faster depends on the types of the inputs. An ad-hoc (ugly) solution might be to check the types and put `CartesianIndices` later.