@vtjnash 2021-04-19
Closing following @kimikage's comment. SCEV should perform that weakening transform. Though perhaps (from the original issue) we need to manually handle `state === nothing`, instead of relying on union splitting?