@kimikage 2021-04-19
However, this PR is not a bad idea from the point of view of readability.